+++ src/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import App from "./home";
-//import Grid from './grid';
+//import App from "./home";
+import Grid from './grid';
 
-ReactDOM.render(<App />, document.getElementById("root"));
+ReactDOM.render(<Grid />, document.getElementById("root"));
